// src/app/modules/product/product.interface.ts
export type TProduct = {
  productName: string;
  productPrice: number;
  productQuantity: number;
  releaseDate: Date;
  brand: string;
  model: string;
  operatingSystem: string;
  storageCapacity: string;
  screenSize: string;
  cameraQuality?: string;
  batteryLife?: string;
  color?: string;
  processor?: string;
  RAM?: string;
};